+++ src/io/io.jl
@@ -323,6 +323,7 @@ function add_system_data!(system_data, surface_normal_method::ColorfieldSurfaceN
     system_data["surface_normal_method"]["ideal_density_threshold"] = surface_normal_method.ideal_density_threshold
     system_data["surface_normal_method"]["interface_taper_start"] = surface_normal_method.interface_taper_start
     system_data["surface_normal_method"]["support_taper_width"] = surface_normal_method.support_taper_width
+    system_data["surface_normal_method"]["normal_smoothing"] = surface_normal_method.normal_smoothing
 end
 
 function add_system_data!(system_data, surface_normal_method::CorrectedCSFSurfaceNormal)